abstract events utils and improve metadata
