+++ app/api/events/[slug]/route.ts
@@ -1,59 +1,5 @@
 import { NextRequest, NextResponse } from 'next/server';
-import fs from 'fs';
-import path from 'path';
-import matter from 'gray-matter';
-import { Event, EventFrontmatter } from '@/app/utils/markdownUtils';
-
-// Define the function to read a specific event file
-function readMarkdownFile(filePath: string): {
-    frontmatter: EventFrontmatter;
-    content: string;
-} {
-    try {
-        const fileContents = fs.readFileSync(filePath, 'utf8');
-        const { data, content } = matter(fileContents);
-
-        return {
-            frontmatter: data as EventFrontmatter,
-            content,
-        };
-    } catch (error) {
-        console.error(`Error reading markdown file: ${filePath}`, error);
-        return {
-            frontmatter: {
-                title: 'Error',
-                description: 'Error loading event',
-                date: new Date().toISOString(),
-                endDate: new Date().toISOString(),
-                image: '/jorbites-social.jpg',
-            },
-            content: 'Error loading event content.',
-        };
-    }
-}
-
-// Get a specific event by slug and language
-function getEventBySlug(slug: string, language: string = 'en'): Event | null {
-    const eventsDirectory = path.join(
-        process.cwd(),
-        'app/content/events',
-        language
-    );
-    const filePath = path.join(eventsDirectory, `${slug}.md`);
-
-    if (!fs.existsSync(filePath)) {
-        return null;
-    }
-
-    const { frontmatter, content } = readMarkdownFile(filePath);
-
-    return {
-        slug,
-        frontmatter,
-        content,
-        language,
-    };
-}
+import { getEventBySlug } from '@/app/utils/event-utils';
 
 interface IParams {
     slug: string;
@@ -76,7 +22,7 @@ export async function GET(
     const language = validLanguages.includes(lang) ? lang : 'en';
 
     // Get the event
-    const event = getEventBySlug(slug, language);
+    const event = await getEventBySlug(slug, language);
 
     if (!event) {
         return NextResponse.json({ error: 'Event not found' }, { status: 404 });
